python/manual-instrumentation/requirements.txt
opentelemetry-exporter-otlp==1.6.2
opentelemetry-exporter-zipkin-json==1.6.2
opentelemetry-sdk==1.6.2
